The sequence of the patient and her family members DNA amplified using methylated and unmethylated primers targeting the A/B region
The sequence of the patient and her family members DNA amplified using methylated and unmethylated primers targeting the A/B region. No change was observed in the amplified sequence of the patient DNA with methylated primers. C is converted to T in the amplified sequence of the patient DNA with unmethylated primers. No change was observed in the amplified sequence of the family members DNA with methylated and unmethylated primers.